build: rework get-all-data to python because NodeJS fetch kinda sucks
