+++ scripts/manual-tests/requirements.txt
@@ -1,0 +1,3 @@
+requests==2.31.0
+autopep8==2.0.4
+pylint==3.0.2
